[WIP] Add some tests for Skills.
