+++ module/lib/skill.mjs
@@ -18,13 +18,6 @@ class Skill {
     }
 
     /**
-     * Validates if the skill with the action and the phase are valid together.
-     */
-    isValid() {
-        throw new Error("Method 'method1()' must be implemented.");
-    }
-
-    /**
      * Processes the action on the skill.
      */
     async process() {
@@ -81,210 +74,10 @@ class Skill {
     }
 }
 
-class CareerFreeSkill extends Skill {
-    constructor(actor, skill, params, options) {
-        super(actor, skill, params, options);
-        this.freeSkillRankAvailable = this.#computeFreeSkillRankAvailable();
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    isValid() {
-        if (!this.isCreation) {
-            return false;
-        }
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    train() {
-        this.skill.rank.careerFree++;
-        this.actor.freeSkillRanks.career.spent++;
-        return this;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-
-    forget() {
-        this.skill.rank.careerFree--;
-        this.actor.freeSkillRanks.career.spent--;
-        return this;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    evaluate() {
-        this.freeSkillRankAvailable = this.#computeFreeSkillRankAvailable();
-        if (this.skill.rank.careerFree < 0) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't forget this rank because it comes from species free bonus!")});
-        }
-
-        this.skill.rank.value = this.skill.rank.base + this.skill.rank.careerFree + this.skill.rank.specializationFree + this.skill.rank.trained
-
-        if (this.skill.rank.value < 0) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't have less than 0 rank!")});
-        }
-
-        if (this.skill.rank.careerFree > 1) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't use more than 1 free skill rank into the same skill!")});
-        }
-
-        if (this.freeSkillRankAvailable < 0) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't use free skill rank anymore. You have used all!")});
-
-        }
-        if (this.freeSkillRankAvailable > 4) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't get more than 4 free skill ranks!")});
-
-        }
-        this.evaluated = true;
-        return this;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    #computeFreeSkillRankAvailable() {
-        return this.actor.freeSkillRanks.career.gained - this.actor.freeSkillRanks.career.spent;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    async updateState() {
-        try {
-            await this.actor.update({'system.progression.freeSkillRanks': this.actor.freeSkillRanks});
-            await this.actor.update({[`system.skills.${this.skill.id}.rank`]: this.skill.rank});
-            return new Promise((resolve, _) => {
-                resolve(this);
-            });
-        } catch (e) {
-            return new Promise((resolve, _) => {
-                resolve(new ErrorSkill(this.actor, this.skill, {}, {message: e.toString()}));
-            });
-        }
-    }
-}
-
-class SpecializationFreeSkill extends Skill {
-    constructor(actor, skill, params, options) {
-        super(actor, skill, params, options);
-        this.freeSkillRankAvailable = this.#computeFreeSkillRankAvailable();
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    isValid() {
-        if (!this.isCreation) {
-            return false;
-        }
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    train() {
-        this.skill.rank.specializationFree++;
-        this.actor.freeSkillRanks.specialization.spent++;
-        return this;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-
-    forget() {
-        this.skill.rank.specializationFree--;
-        this.actor.freeSkillRanks.specialization.spent--;
-        return this;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    evaluate() {
-        this.freeSkillRankAvailable = this.#computeFreeSkillRankAvailable();
-        if (this.skill.rank.specializationFree < 0) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't forget this rank because it comes from species free bonus!")});
-        }
-
-        this.skill.rank.value = this.skill.rank.base + this.skill.rank.careerFree + this.skill.rank.specializationFree + this.skill.rank.trained
-
-        if (this.skill.rank.value < 0) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't have less than 0 rank!")});
-        }
-
-        if (this.skill.rank.specializationFree > 1) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't use more than 1 free skill rank into the same skill!")});
-        }
-
-        if (this.freeSkillRankAvailable < 0) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't use free skill rank anymore. You have used all!")});
-        }
-
-        if (this.freeSkillRankAvailable > 4) {
-            return new ErrorSkill(this.actor, this.skill, {}, {message: ("you can't get more than 4 free skill ranks!")});
-        }
-
-        this.evaluated = true;
-        return this;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    #computeFreeSkillRankAvailable() {
-        return this.actor.freeSkillRanks.specialization.gained - this.actor.freeSkillRanks.specialization.spent;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    async updateState() {
-        try {
-            await this.actor.update({'system.progression.freeSkillRanks': this.actor.freeSkillRanks});
-            await this.actor.update({[`system.skills.${this.skill.id}.rank`]: this.skill.rank});
-            return new Promise((resolve, _) => {
-                resolve(this);
-            });
-        } catch (e) {
-            return new Promise((resolve, _) => {
-                resolve(new ErrorSkill(this.actor, this.skill, {}, {message: e.toString()}));
-            });
-        }
-    }
-}
-
 class TrainedSkill extends Skill {
     constructor(actor, skill, params, options) {
         super(actor, skill, params, options);
         this.#computeFreeSkillRankAvailable();
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
-    isValid() {
-        return true;
     }
 
     /**
@@ -372,14 +165,6 @@ class ErrorSkill extends Skill {
      * @inheritDoc
      * @override
      */
-    isValid() {
-        return false;
-    }
-
-    /**
-     * @inheritDoc
-     * @override
-     */
     process() {
         console.log("Process not implemented.");
     }
@@ -416,8 +201,6 @@ class ErrorSkill extends Skill {
 
 export {
     Skill,
-    CareerFreeSkill,
-    SpecializationFreeSkill,
     TrainedSkill,
     ErrorSkill,
 }
