+++ tests/lib/skills/skill-factory.test.mjs
@@ -2,341 +2,12 @@
 import '../../setupTests.js';
 import {describe, expect, test} from 'vitest'
 import SkillFactory from "../../../module/lib/skills/skill-factory.mjs";
-import {CareerFreeSkill, ErrorSkill, SpecializationFreeSkill, TrainedSkill} from "../../../module/lib/skill.mjs";
+import {ErrorSkill, TrainedSkill} from "../../../module/lib/skill.mjs";
+import {createActor} from "../../utils/actors/actor.mjs";
+import CareerFreeSkill from "../../../module/lib/skills/career-free-skill.mjs";
+import SpecializationFreeSkill from "../../../module/lib/skills/specialization-free-skill.mjs";
 
-/**
- * @returns {SwerpgActor} an actor object
- */
-function createActor() {
-    return {
-        freeSkillRanks: {
-            "career": {
-                "id": "",
-                "name": "",
-                "spent": 0,
-                "gained": 4,
-                "available": 4
-            },
-            "specialization": {
-                "id": "",
-                "name": "",
-                "spent": 0,
-                "gained": 2,
-                "available": 2
-            }
-        },
-        experiencePoints: {
-            "spent": 0,
-            "gained": 0,
-            "startingExperience": 100,
-            "total": 100,
-            "available": 100
-        },
-        system: {
-            skills: {
-                "cool": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "discipline": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "negotiation": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "perception": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "vigilance": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "brawl": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "melee": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "rangedlight": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "rangedheavy": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "gunnery": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "astrogation": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "athletics": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "charm": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "coercion": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "computers": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "coordination": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "deception": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "leadership": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "mechanics": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "medicine": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "pilotingplanetary": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "pilotingspace": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "resilience": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "skulduggery": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "stealth": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "streetwise": {
-                    "rank": {
-                        "base": 1,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 1
-                    }
-                },
-                "survival": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "coreworlds": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "lore": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "outerrim": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "underworld": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "xenology": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                },
-                "education": {
-                    "rank": {
-                        "base": 0,
-                        "careerFree": 0,
-                        "specializationFree": 0,
-                        "trained": 0,
-                        "value": 0
-                    }
-                }
-            }
-        }
-    };
-}
-
-describe("Function build()", () => {
+describe("SkillFactory build()", () => {
     describe("during creation time", () => {
         describe("should create a CareerFreeSkill", () => {
             const expectClassSkill = CareerFreeSkill;
